+++ src/ResourceManager/DataFactoryV2/Commands.DataFactoryV2/Changelog.md
@@ -1,4 +1,4 @@
-<!--
+﻿<!--
     Please leave this section at the top of the change log.
 
     Changes for the current release should go under the section titled "Current Release", and should adhere to the following format:
@@ -19,6 +19,15 @@
 -->
 
 ## Current Release
+* This module is outdated and will go out of support on 29 February 2024.
+* The Az.DataFactory module has all the capabilities of AzureRM.DataFactoryV2 and provides the following improvements:
+    - Greater security with token cache encryption and improved authentication.
+    - Availability in Azure Cloud Shell and on Linux and macOS.
+    - Support for all Azure services.
+    - Allows use of Azure access tokens.
+* We encourage you to start using the Az module as soon as possible to take advantage of these improvements.
+* [Update your scripts](https://aka.ms/azpsmigrate) that use AzureRM PowerShell modules to use Az PowerShell modules by 29 February 2024.
+* To automatically update your scripts, follow the [quickstart guide](https://aka.ms/azpsmigratequick).
 
 ## Version 0.5.11
 * Updated the ADF .Net SDK version to 2.3.0.